verticalTreversal
